@@ -1,15 +1,6 @@
 <template>
-  <div id="app">
-    <router-view />
-  </div>
+  <router-view />
 </template>
-
-<script>
-export default {
-  name: "App",
-};
-</script>
-
 
 <style lang="css">
 @import url("https://rsms.me/inter/inter.css");
@@ -131,4 +122,3 @@ a {
   color: inherit;
 }
 </style>
-
