@@ -1,5 +1,5 @@
 export const COMETCHAT_CONSTANTS = {
-  APP_ID: "APP_ID",
-  REGION: "REGION",
-  AUTH_KEY: "AUTH_KEY",
+  APP_ID: "",
+  REGION: "",
+  AUTH_KEY: "",
 };
